feat: Add username to local storage after successful login
Update the login API to store the username in the local storage along with the token. This will allow easier access to the username throughout the application.

Update the login API to store the username in the local storage along with the token. This will allow easier access to the username throughout the application.